Refactor multisort unit tests.
