+++ realm/src/androidTest/java/io/realm/RealmTest.java
@@ -339,76 +339,6 @@ public class RealmTest extends AndroidTestCase {
         } catch (IllegalArgumentException ignored) {}
     }
 
-    public void testSortTwoFields() {
-        io.realm.TestHelper.populateForMultiSort(testRealm);
-
-        RealmResults<AllTypes> results1 = testRealm.allObjectsSorted(AllTypes.class,
-                new String[]{FIELD_STRING, FIELD_LONG},
-                new boolean[] {RealmResults.SORT_ORDER_ASCENDING, RealmResults.SORT_ORDER_ASCENDING});
-
-        assertEquals(3, results1.size());
-
-        assertEquals("Adam", results1.get(0).getColumnString());
-        assertEquals(4, results1.get(0).getColumnLong());
-
-        assertEquals("Adam", results1.get(1).getColumnString());
-        assertEquals(5, results1.get(1).getColumnLong());
-
-        assertEquals("Brian", results1.get(2).getColumnString());
-        assertEquals(4, results1.get(2).getColumnLong());
-
-        RealmResults<AllTypes> results2 = testRealm.allObjectsSorted(AllTypes.class,
-                new String[]{FIELD_LONG, FIELD_STRING},
-                new boolean[]{RealmResults.SORT_ORDER_ASCENDING, RealmResults.SORT_ORDER_ASCENDING});
-
-        assertEquals(3, results2.size());
-
-        assertEquals("Adam", results2.get(0).getColumnString());
-        assertEquals(4, results2.get(0).getColumnLong());
-
-        assertEquals("Brian", results2.get(1).getColumnString());
-        assertEquals(4, results2.get(1).getColumnLong());
-
-        assertEquals("Adam", results2.get(2).getColumnString());
-        assertEquals(5, results2.get(2).getColumnLong());
-    }
-
-    public void testSortMultiFailures() {
-        RealmResults<AllTypes> allTypes = testRealm.allObjects(AllTypes.class);
-
-        // zero fields specified
-        try {
-            testRealm.allObjectsSorted(AllTypes.class, new String[]{}, new boolean[]{});
-            fail();
-        } catch (IllegalArgumentException ignored) {}
-
-        // number of fields and sorting orders don't match
-        try {
-            testRealm.allObjectsSorted(AllTypes.class,
-                    new String[]{FIELD_STRING},
-                    new boolean[]{RealmResults.SORT_ORDER_ASCENDING, RealmResults.SORT_ORDER_ASCENDING});
-            fail();
-        } catch (IllegalArgumentException ignored) {}
-
-        // null is not allowed
-        try {
-            testRealm.allObjectsSorted(AllTypes.class, null, null);
-            fail();
-        } catch (IllegalArgumentException ignored) {}
-        try {
-            testRealm.allObjectsSorted(AllTypes.class, new String[]{FIELD_STRING}, null);
-            fail();
-        } catch (IllegalArgumentException ignored) {}
-
-        // non-existing field name
-        try {
-            testRealm.allObjectsSorted(AllTypes.class,
-                    new String[]{FIELD_STRING, "dont-exist"},
-                    new boolean[]{RealmResults.SORT_ORDER_ASCENDING, RealmResults.SORT_ORDER_ASCENDING});
-            fail();
-        } catch (IllegalArgumentException ignored) {}
-    }
-
     public void testSortSingleField() {
         populateTestRealm();
         RealmResults<AllTypes> sortedList = testRealm.allObjectsSorted(AllTypes.class,
